+++ packages/client/src/shared/ui/index.ts
@@ -2,4 +2,4 @@ export { AuthForm } from './auth-form'
 export { useError } from './use-error'
 export { AppLayout } from './app-layout'
 export { Logo } from './Logo'
-export { InputField } from './fields/input-field'
+export { InputField } from './input/input-field'
