+++ packages/client/src/shared/ui/input/config.ts
